Update layout.tsx
i have not found our head section of the page so i added the link for tittlebar icon in the main section.please consider to replace it where the heaad section you add. this change make our website more interactive and without it. when i stay at another tab of browser its like our shefra webpage is offline or not responding.please consider it brother